Actually rename packages
