+++ src/main/java/org/cryptimeleon/math/structures/groups/elliptic/type3/mcl/MclBilinearGroup.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.math.pairings.mcl;
+package org.cryptimeleon.math.structures.groups.elliptic.type3.mcl;
 
 import org.cryptimeleon.math.serialization.Representation;
 import org.cryptimeleon.math.structures.groups.lazy.LazyBilinearGroup;
