+++ src/test/java/org/cryptimeleon/math/structures/groups/elliptic/type3/mcl/GroupTest.java
@@ -1,8 +1,8 @@
-package org.cryptimeleon.math.pairings.mcl.test;
+package org.cryptimeleon.math.structures.groups.elliptic.type3.mcl;
 
-import org.cryptimeleon.math.pairings.mcl.MclGroup1Impl;
-import org.cryptimeleon.math.pairings.mcl.MclGroup2Impl;
-import org.cryptimeleon.math.pairings.mcl.MclGroupTImpl;
+import org.cryptimeleon.math.structures.groups.elliptic.type3.mcl.MclGroup1Impl;
+import org.cryptimeleon.math.structures.groups.elliptic.type3.mcl.MclGroup2Impl;
+import org.cryptimeleon.math.structures.groups.elliptic.type3.mcl.MclGroupTImpl;
 import org.cryptimeleon.math.structures.GroupTests;
 import org.cryptimeleon.math.structures.groups.basic.BasicGroup;
 import org.junit.runner.RunWith;
