+++ src/test/java/org/cryptimeleon/math/structures/groups/elliptic/type3/mcl/PairingTest.java
@@ -1,7 +1,7 @@
-package org.cryptimeleon.math.pairings.mcl.test;
+package org.cryptimeleon.math.structures.groups.elliptic.type3.mcl;
 
 import org.cryptimeleon.math.pairings.PairingTests;
-import org.cryptimeleon.math.pairings.mcl.MclBilinearGroup;
+import org.cryptimeleon.math.structures.groups.elliptic.type3.mcl.MclBilinearGroup;
 import org.cryptimeleon.math.structures.groups.elliptic.BilinearMap;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
